Fix bugs with ternary operator, fix operator support in function arguments, Fix some design problems.
Now ternary operator works for chain cases , registration of ternar operator
is done with a single addOperator.
For ternary operator correct work we assume that '?' is a separator , and
':' is operator which take 3 arguments , in infixToPostfix we hold flag and
count of ternarOperator symbol.
For operator support in function arguments added comma parsing in infixToPostfix

operatorSymbols moved to Lexer , isFunction and isOperator also.